+++ src/cli/commands/env.ts
@@ -1,5 +1,5 @@
 import { Command } from 'commander';
-import { ConfigOption, configOption } from '../options';
+import { ConfigOption } from '../options';
 import { Logger } from '../../utils/logger';
 import { getPackageVersion } from '../../utils';
 import { resolveConfigFile } from '../../playwright/loadConfig';
@@ -9,7 +9,7 @@ const logger = new Logger({ verbose: true });
 
 export const envCommand = new Command('env')
   .description('Prints environment info')
-  .addOption(configOption)
+  .configureHelp({ showGlobalOptions: true })
   .action((opts: ConfigOption) => {
     logger.log(`Playwright-bdd environment info:\n`);
     logger.log(`platform: ${process.platform}`);