+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Dev
+## 7.2.2
 * fix: provide empty world for Playwright-style steps ([#208](https://github.com/vitalets/playwright-bdd/issues/208))
 * fix: improve CLI output ([#206](https://github.com/vitalets/playwright-bdd/issues/206))